requirements.txt
grpc==1.0.0
grpcio==1.46.3
lap==0.4.0
numpy==1.22.4
nvidia_ml_py3==7.352.0
nvidia_tensorrt==8.4.1.5
onnx==1.11.0
onnxruntime==1.12.1
onnxruntime_gpu==1.11.1
opencv_python==4.5.1.48
Pillow==9.2.0
protobuf==4.21.7
pycuda==2020.1
pynvml==11.4.1
PyYAML==6.0
scipy==1.6.2
tensorrt==0.0.1.dev5
torch==1.11.0+cu113
torchvision==0.12.0+cu113
